Update Omega-3 section in README
The Omega-3 section in the README was rewritten for clarity and precision. It now specifically highlights the importance of EPA & DHA and underscores the necessity for individuals with the APOE-4 SNP to prioritize Omega-3 intake from fish. The significance of an Omega-3 Index of 8–11% for overall health maintenance is also emphasized.